requirements.txt
# Web framework
Flask==2.3.3
Werkzeug==2.3.7

# Core deep learning framework
torch>=1.12.0
torchvision>=0.13.0

# Data processing and scientific computing
numpy>=1.21.0
scikit-learn>=1.0.0

# Image processing and geospatial
rasterio>=1.2.0
albumentations>=1.2.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Progress bars and utilities
tqdm>=4.64.0

# Model summary and utilities
torchsummary>=1.5.1

# Additional utilities
opencv-python>=4.5.0

# Development and testing (optional)
jupyter>=1.0.0
ipykernel>=6.0.0 